Finished MyProfile V1
Added tabs to MyProfile, one to display user settings and info, and another to update user settings and info, including changing the profile picture, displayname, and toggle beginner mode. I also set home and log in screens to render different html based on whether the user is logged in; this fixes FOUC problems where the system has to figure out if the user is logged in after render, then update what is shown. Now, it checks the state of the user before rendering any html