Create League and League View Modal Created
Created a function and interface to create a new league, including roster settings, scoring settings, and inviting other users. Leagues appear on the home page for users, displaying a card for each league a user is included as a member of. Clicking into a league will pull up a modal with different tabs for the league where one can see things like My team, matchup, add players, league chat, and league settings.

Also added the logo as a browser tab icon and added the logo to the nav bar