Add Players 1.0, MyTeam 1.0, and Draft 0.5 complete
From the last couple of days:

Finished up the AddFreeAgents page to allow users to add free agents until their roster is full. Need to add restrictions based on the draft status.

My Team was quick, just displays all the players in the current users team, with a button to drop the player.

Draft went suprisingly smoothly, reusing a lot of code from AddFreeAgents, except there are a couple of restrictions to only allow you to draft if it is your turn to draft. Need to add new fields to league creation to make draft usable in all leagues