Friend Requests, Friend Chats, and League Invitation
Over the past few days:

Added A friend request system and a chat system for friends to direct message in real time.

Added league invite requests for the createLeague screen. You can now select from your friends list or type in the email address of someone you want to add to the league. Invites appear on the home page and can be accepted or declined. Once a league is full, all leftover invitations are deleted.