Many Small Changes
Added lots of small changes to many pages, such as restricting free agency additions till after the draft is over. Making duplicate invites to leagues not send, making league draft order update dynamically for all users, moving the log in/sign up screen to a modal that can be opened from the nav bar when not signed in.