Nav Bar Updates
Updated nav bar to display the profile picture for the user, with a dropdown menu to navigate to different pages.
Made a password reset page that can be used from the main screen.
Moved the log in to a modal rather than its own page.
and updated some styling on the league cards on the home page