Quick Styling fixes
Quick fixes on minor styling elements missed in the last update, such as outlines on the dropdown toggle and click effects. Fixed issue with league cards displaying weirdly when a user only has 1 or 2 leagues total. Added some filler content for the home page when no user is signed in--will flesh out later