Added HuddleHeroes images
Added images for MyHuddleHero. Edited to show a random huddle hero on a page with the feature unless a specific hero is specified in the component call

Began styling the HuddleHero icon so it takes up less space.

Added a My Huddle Hero explanation on the register user page, showing all the HuddleHeroes as well as a brief description of the feature for users