My Huddle Hero icon placement, fixes to My Team
Fixed My Team to show players with no Data. It was causing crashes if a player had no data for a specific week.

Set MyHuddleHero to show in the Modal Title for each league, using the active tab as a key to what data to show inside. Also created components for each tutorial to live in to allow easier access/editing later